+++ src/app/components/operator-home/operator-home.component.ts
@@ -5,6 +5,9 @@ import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import { TokenStorageService } from 'src/app/_services/token-storage.service';
+import { SpinnerService } from '../../_services/spinner.service';
+import Swal from 'sweetalert2';
+import { faPencilAlt, faTrash } from '@fortawesome/free-solid-svg-icons';
 
 const READ_OPERATOR_API = 'http://10.1.137.50:8760/admin/v1/getAll?role=operator'
 const DELETE_OPERATOR_API = 'http://10.1.137.50:8760/admin/v1/delete/'
@@ -16,6 +19,9 @@ const GET_OPERATOR_BYID_API = 'http://10.1.137.50:8080/auth/user/v1/'
   styleUrls: ['./operator-home.component.css']
 })
 export class OperatorHomeComponent implements OnInit {
+
+  faEdit = faPencilAlt;
+  faDelete = faTrash;
 
   // token for get anything data
   httpOptions_base = {
@@ -34,10 +40,12 @@ export class OperatorHomeComponent implements OnInit {
   constructor(
     private http: HttpClient,
     private router: Router,
-    private token: TokenStorageService
+    private token: TokenStorageService,
+    private spiner: SpinnerService
   ) { }
 
   ngOnInit(): void {
+    this.spiner.isLoading = true
     this.getAllData()
   }
 
@@ -56,25 +64,52 @@ export class OperatorHomeComponent implements OnInit {
         this.dataSource = new MatTableDataSource(isi.content)
         this.dataSource.paginator = this.paginator
         this.dataSource.sort = this.sort
+        this.spiner.isLoading = false
       },
         err => {
-          alert(err.error.message)
+          this.spiner.isLoading = false
+          Swal.fire({
+            title: 'Error',
+            text: err.error.message,
+            icon: 'error'
+          })
           console.log(err)
         }
       )
   }
 
   deleteOperator(id: string) {
-    this.http.delete<any>(DELETE_OPERATOR_API + id, this.httpOptions_base)
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.http.delete<any>(DELETE_OPERATOR_API + id, this.httpOptions_base)
       .subscribe(isi => {
-        
+
         this.getAllData()
       },
         err => {
-          alert(err.error.message)
+          Swal.fire({
+            title: 'Error',
+            text: err.error.message,
+            icon: 'error'
+          })
           console.log(err)
         }
       )
+      } else {
+        Swal.fire(
+          'Cancelled',
+          'Your operator is safe :)',
+          'error'
+        )
+      }
+    })
   }
-
 }
