+++ src/app/components/penjual-home/penjual-home.component.ts
@@ -3,8 +3,11 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
+import { faPencilAlt, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { LelangDataService } from 'src/app/_services/lelang-data.service';
 import { PenjualDataService } from 'src/app/_services/penjual-data.service';
+import { SpinnerService } from 'src/app/_services/spinner.service';
+import Swal from 'sweetalert2';
 
 export interface Seller {
   id: string,
@@ -30,7 +33,8 @@ export interface Seller {
   styleUrls: ['./penjual-home.component.css']
 })
 export class PenjualHomeComponent implements OnInit {
-
+  faEdit = faPencilAlt;
+  faDelete = faTrash;
   sellers: Seller[] = [];
   displayedColumns: string[] = [ 'name', 'address', 'industry', 'phoneNum', 'fax', 'action'];
   dataSource = new MatTableDataSource(this.sellers);
@@ -39,15 +43,34 @@ export class PenjualHomeComponent implements OnInit {
   @ViewChild(MatSort) sort!: MatSort;
 
   deleteSeller(id: string) {
-    
-    this.sellerserv.deleteSeller(id).subscribe(()=>{
-      this.getSeller()
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.sellerserv.deleteSeller(id).subscribe(()=>{
+          this.getSeller()
+        })
+      } else {
+        Swal.fire(
+          'Cancelled',
+          'Your seller file is safe :)',
+          'error'
+        )
+      }
     })
   }
-  constructor(private sellerserv:PenjualDataService, private router:Router) {
-  }
+
+  constructor(private sellerserv:PenjualDataService, private router:Router, private spinner: SpinnerService) { }
+
 
   ngOnInit(): void {
+    this.spinner.isLoading = true
     this.getSeller()
   }
 
@@ -58,6 +81,7 @@ export class PenjualHomeComponent implements OnInit {
       this.dataSource = new MatTableDataSource(this.sellers)
       this.dataSource.paginator = this.paginator;
       this.dataSource.sort = this.sort;
+      this.spinner.isLoading = false
       // console.log(this.sellers)
     })
   }
